refactor: 네트워크 에러관리 폼으로 이동
