+++ src/features/registerCard/model/registerCardForm.ts
@@ -21,6 +21,11 @@ export interface FieldData {
   password: string;
 }
 
+export type ServerError = {
+  field: ServerErrorField;
+  message: string;
+} | null;
+
 export type ServerErrorField = 'numbers' | 'expiryDate' | 'cvc';
 export type ServerFieldErrors = Partial<Record<ServerErrorField, string>>;
 export const SERVER_ERROR_FIELD_MAP: Record<RegisterCardErrorCode, ServerErrorField> = {
